updated multiple files and display functions and others
