+++ README.md
@@ -5,10 +5,13 @@
 Gophers Dataframe Library (not for Pandas or Polar Bears; Gophers only)
 
 - Go Package
+-- go get https://www.github.com/speartech/gophers.git
 -- go run .\main.go
 
 - Python Package
 -- python -m pip install gophers
+-- import gophers as g
+-- df = g.ReadJSON("<JSON String or Filepath>")
 
 - NPM Package
--- TBD
+-- npm i gophers OR by CDN via import { Gophers } from 'https://unpkg.com/gophers@0.0.1/gophers.js';
