+++ go/gophers/dashboard.go
@@ -1,0 +1,43 @@
+package gophers
+
+// main dashboard object for adding html pages, charts, and inputs for a single html output
+type Dashboard struct {
+	pages map[string]interface{}
+}
+
+// create page
+func (dash *Dashboard) AddPage(name string) {
+
+}
+
+// add text input
+
+// add slider
+
+// add dropdown (array of selections)
+
+// add chart
+
+// add title
+
+// add header1
+
+// add header2
+
+// add header3
+
+// add header4
+
+// add normal text (paragraph)
+
+// add button
+
+// set primary color
+
+// set secondary color
+
+// set tertiary color
+
+// set neutral color
+
+//
