+++ go/main.go
@@ -1,7 +1,8 @@
 package main
 
 import (
-	gf "gophers/gophers"
+	"fmt"
+	g "gophers/gophers"
 )
 
 func main() {
@@ -14,13 +15,13 @@ func main() {
 	jsonStr := "test_data/jsontest.json"
 	// Create a DataFrame from the unmarshalled rows.
 	// df := gf.ReadNDJSON(jsonStr)
-	df := gf.ReadJSON(jsonStr)
+	df := g.ReadJSON(jsonStr)
 	// df := gf.ReadCSV(jsonStr)
 
 	// Print the DataFrame.
 	// df.Show(7, 1)
 
-	// df.Head(5)
+	df.Head(5)
 
 	// df.Tail(8)
 
@@ -33,7 +34,12 @@ func main() {
 
 	// df = df.Select("age", "score")
 
-	df = df.Rename("age", "agey")
+	df = df.Rename("namythingamajigya", "name")
+	df = df.Column("name2", g.Values("name"))
+	data := df.ToJSON()
+	fmt.Println(data)
+	// df = df.FillNA("tyler")
+	df.BrowserDisplay()
 	// df = df.DropNA()
 	df.Show(10)
 }
