+++ go/test_data/csvtest.csv
@@ -1,3 +1,3 @@
 namythingamajigya,age,score
 Alicandro,30,85.5
-Bob,25,90.0
+Bob,,90.0
