+++ go/test_data/jsontest.json
@@ -1,4 +1,4 @@
 [
     {"namythingamajigya": "Alicandro", "age": "30", "score": 85.5},
-    {"namythingamajigya": "Bob", "age": "25", "score": 90.0}
+    {"namythingamajigya": "Bob", "age": "", "score": 90.0}
 ]
